This is just the static code to build a chart
The code shows a .gs file which is on the server side, and an html file which is on the client side. It is static right now, in the sense that the numbers are hard coded into the formulas/functions in the .gs file. The next step will be to get the numbers extracted directly from the google sheet, but that currently is hitting timing snags.